fix: Keep ExternalObject from merging with themselves
The real fix in this is only the line in lookup_or_create that checks if
the provided external_object_id isn't the same as the one found while
resolving the links.

The rest is about correctly using savepoints (begin_nested) and making
intelligent batch updates instead of python loops.